テストユーザーログイン改善
